ai_models/biometric_deepface/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
# Nota: opencv, numpy y face_recognition NO van aquí 
# porque usaremos la imagen base que ya los trae instalados.